Hacks to make spike behave like the verilog cpu
The original code basically loads the elf file directly into ram via an
emulated debug interface, whereas the verilog cpu expects the code to be
in rom with ram uninitialized. Since the elf loader is subject to the
same memory access controls as the cpu, we use a "fake rom" that is
writable during the elf loading phase but read-only once execution
starts. We also make a hard call to exit() when the magic test
instruction is encountered since spike also relies on the semihosting
interface to exit normally. In the future the verilog cpu will probably
also have a real debug/semihosting interface in which case this hack can
be changed to something more reasonable.